requirements.txt
about-time==4.2.1
alive-progress==3.3.0
amqp==5.3.1
asgiref==3.9.1
attrs==25.4.0
billiard==4.2.1
black==25.1.0
celery==5.5.3
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==5.0.8
django-ckeditor-5==0.2.18
django-cors-headers==4.8.0
django-default==0.0.9
django-filter==25.1
django-modeltranslation==0.19.16
django-redis==6.0.0
django-rosetta==0.10.2
django-unfold==0.65.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.9.1
Faker==37.6.0
flower==2.0.1
graphemeu==0.7.2
gunicorn==23.0.0
humanize==4.13.0
idna==3.10
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.5.4
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
platformdirs==4.4.0
polib==1.2.0
prometheus_client==0.22.1
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pydot==4.0.1
PyJWT==2.10.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
redis==6.4.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
tornado==6.5.2
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.13
